Refactored repository folder
