+++ src/repository/index.ts
@@ -1,6 +1,6 @@
 import { DatabaseConfig } from '@infrastructure/config/index.js';
-import Database from '@repository/postgres/database.js';
-import NoteStorage from '@repository/postgres/orm/note.js';
+import DatabaseSequelize from '@repository/storage/postgres/database.sequlieze.js';
+import NoteStorage from '@repository/storage/postgres/orm/note.js';
 import NoteRepository from '@repository/note.js';
 
 /**
@@ -19,7 +19,7 @@ export interface Repositories {
  * @param databaseConfig - database config
  */
 export async function initRepositories(databaseConfig: DatabaseConfig): Promise<Repositories> {
-  const database = new Database(databaseConfig);
+  const database = new DatabaseSequelize(databaseConfig);
 
   const conn = await database.connect();
